desktop-windows/requirements.txt
aiohttp>=3.10,<4
Pillow>=10,<12
numpy>=2.0,<3
pyvirtualcam>=0.12,<0.15
